requirements.txt
openai
openai-whisper
opencv-python
ffmpeg-python
fastapi
uvicorn
imagehash
pillow
